LoggingCommands.cs Added
Moves logging commands to this class from Administration.cs, it was getting cluttered.
