+++ Kaguya/Kaguya.csproj
@@ -310,6 +310,8 @@
     <Compile Include="Core\PlayData.cs" />
     <Compile Include="Modules\Administration\Administration.cs" />
     <Compile Include="Modules\Administration\AutoModerator.cs" />
+    <Compile Include="Modules\Administration\LoggingCommands.cs" />
+    <Compile Include="Modules\Administration\Shadowbans.cs" />
     <Compile Include="Modules\Currency\Currency.cs" />
     <Compile Include="Modules\EXP\EXP.cs" />
     <Compile Include="Modules\Fun\Fun.cs" />
