+++ Kaguya/Modules/Administration/Administration.cs
@@ -8,6 +8,7 @@ using Kaguya.Core.CommandHandler;
 using Kaguya.Core.Commands;
 using Kaguya.Core.Server_Files;
 using Kaguya.Core.UserAccounts;
+using Kaguya.Modules.Administration;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -863,52 +864,6 @@ namespace Kaguya.Modules
             logger.ConsoleGuildAdvisory(Context.Guild, "KaguyaExit command executed.");
         }
 
-        [Command("kick")] //admin
-        [Alias("k")]
-        [RequireUserPermission(GuildPermission.KickMembers)]
-        [RequireBotPermission(GuildPermission.KickMembers)]
-        public async Task Kick(IGuildUser user, string reason = "No reason provided.")
-        {
-            stopWatch.Start();
-
-            var sOwner = Context.Guild.Owner;
-
-            if (user.Id != sOwner.Id)
-            {
-                if (reason != "No reason provided.")
-                {
-                    await user.KickAsync(reason);
-                    embed.WithTitle($"User Kicked");
-                    embed.WithDescription($"`{Context.User}` has kicked `{user}` with reason: \"{reason}\"");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
-                }
-                else
-                {
-                    await user.KickAsync(reason);
-                    embed.WithTitle($"User Kicked");
-                    embed.WithDescription($"`{Context.User.Mention}` has kicked `{user}` without a specified reason.");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
-                }
-            }
-
-            else if (user.Id == Context.User.Id)
-            {
-                embed.WithDescription($"**{Context.User.Mention} You may not ban yourself!**");
-                embed.WithColor(Red);
-                await BE();
-            }
-            else if (user.Id == sOwner.Id)
-            {
-                embed.WithDescription($"**{Context.User.Mention} I cannot ban the server owner!**");
-                embed.WithColor(Red);
-                await BE();
-            }
-        }
-
         [Command("ban")] //admin
         [Alias("b")]
         [RequireUserPermission(GuildPermission.BanMembers)]
@@ -979,6 +934,52 @@ namespace Kaguya.Modules
             logger.ConsoleGuildAdvisory("Users massbanned.");
         }
 
+        [Command("kick")] //admin
+        [Alias("k")]
+        [RequireUserPermission(GuildPermission.KickMembers)]
+        [RequireBotPermission(GuildPermission.KickMembers)]
+        public async Task Kick(IGuildUser user, string reason = "No reason provided.")
+        {
+            stopWatch.Start();
+
+            var sOwner = Context.Guild.Owner;
+
+            if (user.Id != sOwner.Id)
+            {
+                if (reason != "No reason provided.")
+                {
+                    await user.KickAsync(reason);
+                    embed.WithTitle($"User Kicked");
+                    embed.WithDescription($"`{Context.User}` has kicked `{user}` with reason: \"{reason}\"");
+                    embed.WithColor(Pink);
+                    await BE(); stopWatch.Stop();
+                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
+                }
+                else
+                {
+                    await user.KickAsync(reason);
+                    embed.WithTitle($"User Kicked");
+                    embed.WithDescription($"`{Context.User.Mention}` has kicked `{user}` without a specified reason.");
+                    embed.WithColor(Pink);
+                    await BE(); stopWatch.Stop();
+                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
+                }
+            }
+
+            else if (user.Id == Context.User.Id)
+            {
+                embed.WithDescription($"**{Context.User.Mention} You may not ban yourself!**");
+                embed.WithColor(Red);
+                await BE();
+            }
+            else if (user.Id == sOwner.Id)
+            {
+                embed.WithDescription($"**{Context.User.Mention} I cannot ban the server owner!**");
+                embed.WithColor(Red);
+                await BE();
+            }
+        }
+
         [Command("masskick")] //administration
         [RequireUserPermission(GuildPermission.Administrator)]
         [RequireBotPermission(GuildPermission.Administrator)]
@@ -1012,6 +1013,44 @@ namespace Kaguya.Modules
             logger.ConsoleGuildAdvisory(Context.Guild, "Users masskicked.");
         }
 
+        [Command("clear")] //administration
+        [Alias("c", "purge")]
+        [RequireUserPermission(GuildPermission.ManageMessages)]
+        [RequireBotPermission(GuildPermission.ManageMessages)]
+        public async Task ClearMessages(int amount = 10)
+        {
+            stopWatch.Start();
+
+            if(!(amount > 0))
+            {
+                await GlobalCommandResponses.CreateCommandError(Context,
+                    stopWatch.ElapsedMilliseconds,
+                    CommandError.Unsuccessful, 
+                    "User tried to clear less than one message.",
+                    "Clearing Messages", 
+                    "The number of messages to be deleted must be greater than zero!");
+                return;
+            }
+
+            if(!(amount <= 100))
+            {
+                await GlobalCommandResponses.CreateCommandError(Context,
+                    stopWatch.ElapsedMilliseconds,
+                    CommandError.Unsuccessful,
+                    "User tried to clear more than 100 messages.",
+                    "Clearing Messages",
+                    "The number of messages to be deleted must not be more than 100!");
+                return;
+            }
+
+            var messages = await Context.Channel.GetMessagesAsync(amount + 1).FlattenAsync();
+            await (Context.Channel as SocketTextChannel).DeleteMessagesAsync(messages);
+            var m = await ReplyAsync($"Clearing of messages completed. This message will be deleted in 3 seconds.");
+            await m.DeleteAsync();
+            logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
+            stopWatch.Stop();
+        }
+
         [Command("shadowban")]
         [RequireUserPermission(GuildPermission.BanMembers)]
         [RequireBotPermission(GuildPermission.BanMembers)]
@@ -1029,7 +1068,7 @@ namespace Kaguya.Modules
             var channels = Context.Guild.Channels;
             int i = 0;
 
-            foreach(var channel in channels)
+            foreach (var channel in channels)
             {
                 i++;
                 await channel.AddPermissionOverwriteAsync(user, OverwritePermissions.DenyAll(channel));
@@ -1079,374 +1118,11 @@ namespace Kaguya.Modules
             logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
             logger.ConsoleGuildAdvisory(Context.Guild, user as SocketGuildUser, stopWatch.ElapsedMilliseconds, "User un-shadowbanned.");
 
-            if(server.LogUnShadowbans != 0)
+            if (server.LogUnShadowbans != 0)
             {
                 KaguyaLogMethods logMethods = new KaguyaLogMethods();
                 await logMethods.LoggingUserUnShadowbanned(user as SocketUser, Context.Guild);
             }
         }
-
-        [Command("clear")] //administration
-        [Alias("c", "purge")]
-        [RequireUserPermission(GuildPermission.ManageMessages)]
-        [RequireBotPermission(GuildPermission.ManageMessages)]
-        public async Task ClearMessages(int amount = 10)
-        {
-            stopWatch.Start();
-
-            if(!(amount > 0))
-            {
-                await GlobalCommandResponses.CreateCommandError(Context,
-                    stopWatch.ElapsedMilliseconds,
-                    CommandError.Unsuccessful, 
-                    "User tried to clear less than one message.",
-                    "Clearing Messages", 
-                    "The number of messages to be deleted must be greater than zero!");
-                return;
-            }
-
-            if(!(amount <= 100))
-            {
-                await GlobalCommandResponses.CreateCommandError(Context,
-                    stopWatch.ElapsedMilliseconds,
-                    CommandError.Unsuccessful,
-                    "User tried to clear more than 100 messages.",
-                    "Clearing Messages",
-                    "The number of messages to be deleted must not be more than 100!");
-                return;
-            }
-
-            var messages = await Context.Channel.GetMessagesAsync(amount + 1).FlattenAsync();
-            await (Context.Channel as SocketTextChannel).DeleteMessagesAsync(messages);
-            var m = await ReplyAsync($"Clearing of messages completed. This message will be deleted in 3 seconds.");
-            await m.DeleteAsync();
-            logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
-            stopWatch.Stop();
-        }
-
-        [Command("setlogchannel")] //administration
-        [Alias("log")]
-        [RequireUserPermission(GuildPermission.Administrator)]
-        [RequireBotPermission(GuildPermission.Administrator)]
-        public async Task SetLogChannel(string logType, IGuildChannel channel)
-        {
-            stopWatch.Start();
-            var server = Servers.GetServer(Context.Guild);
-            ulong logChannelID = channel.Id;
-
-            switch (logType.ToLower())
-            {
-                case "deletedmessages":
-                    server.LogDeletedMessages = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `Deleted Messages` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "updatedmessages":
-                    server.LogUpdatedMessages = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `Edited Messages` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userjoins":
-                    server.LogWhenUserJoins = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `User Joins` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userleaves":
-                    server.LogWhenUserLeaves = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `User Leaves` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "bans":
-                    server.LogBans = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `Bans` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "unbans":
-                    server.LogUnbans = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `Unbans` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "changestologsettings":
-                    server.LogChangesToLogSettings = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `changes to log settings` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "filteredphrases":
-                    server.LogWhenUserSaysFilteredPhrase = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `Filtered Phrases` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userconnectstovoice":
-                    server.LogWhenUserConnectsToVoiceChannel = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `user connected to voice` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userdisconnectsfromvoice":
-                    server.LogWhenUserDisconnectsFromVoiceChannel = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages for `user disconnected from voice` will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "levelannouncements":
-                    server.LogLevelUpAnnouncements = logChannelID;
-                    Servers.SaveServers();
-                    await GlobalCommandResponses.CreateCommandResponse(Context,
-                        stopWatch.ElapsedMilliseconds,
-                        "Log Channel Set",
-                        $"{Context.User.Mention} All log messages for `level announcements` will be sent in channel {channel.Name}");
-                    break;
-                case "shadowbans":
-                    server.LogShadowbans = logChannelID;
-                    Servers.SaveServers();
-                    await GlobalCommandResponses.CreateCommandResponse(Context,
-                        stopWatch.ElapsedMilliseconds,
-                        "Log Channel Set",
-                        $"{Context.User.Mention} All log messages for `shadowbans` will be sent in channel {channel.Name}");
-                    break;
-                case "unshadowbans":
-                    server.LogUnShadowbans = logChannelID;
-                    Servers.SaveServers();
-                    await GlobalCommandResponses.CreateCommandResponse(Context,
-                        stopWatch.ElapsedMilliseconds,
-                        "Log Channel Set",
-                        $"{Context.User.Mention} All log messages for `un-shadowbans` will be sent in channel {channel.Name}");
-                    break;
-                case "all":
-                    server.LogDeletedMessages = logChannelID;
-                    server.LogUpdatedMessages = logChannelID;
-                    server.LogWhenUserJoins = logChannelID;
-                    server.LogWhenUserLeaves = logChannelID;
-                    server.LogBans = logChannelID;
-                    server.LogUnbans = logChannelID;
-                    server.LogChangesToLogSettings = logChannelID;
-                    server.LogWhenUserSaysFilteredPhrase = logChannelID;
-                    server.LogWhenUserConnectsToVoiceChannel = logChannelID;
-                    server.LogWhenUserDisconnectsFromVoiceChannel = logChannelID;
-                    server.LogLevelUpAnnouncements = logChannelID;
-                    server.LogShadowbans = logChannelID;
-                    server.LogUnShadowbans = logChannelID;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} All log messages will be sent in channel {channel.Name}");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                default:
-                    embed.WithTitle("Invalid Log Specification");
-                    embed.WithDescription($"{Context.User.Mention} Invalid logging type!");
-                    embed.WithColor(Red);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-            }
-        }
-
-        [Command("resetlogchannel")] //administration
-        [Alias("rlog")]
-        [RequireUserPermission(GuildPermission.Administrator)]
-        [RequireBotPermission(GuildPermission.Administrator)]
-        public async Task ResetLogChannel(string logType)
-        {
-            stopWatch.Start();
-            var server = Servers.GetServer(Context.Guild);
-
-            switch (logType.ToLower())
-            {
-                case "deletedmessages":
-                    server.LogDeletedMessages = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `Deleted Messages` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "updatedmessages":
-                    server.LogUpdatedMessages = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `Updated Messages` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userjoins":
-                    server.LogWhenUserJoins = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `User Joins` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userleaves":
-                    server.LogWhenUserLeaves = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `User Leaves` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "bans":
-                    server.LogBans = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `Bans` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "unbans":
-                    server.LogUnbans = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `Unbans` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "changestologsettings":
-                    server.LogChangesToLogSettings = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `changes to log settings` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "filteredphrases":
-                    server.LogWhenUserSaysFilteredPhrase = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `Filtered Phrases` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userconnectstovoice":
-                    server.LogWhenUserConnectsToVoiceChannel = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `user connects to voice` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "userdisconnectsfromvoice":
-                    server.LogWhenUserDisconnectsFromVoiceChannel = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Set");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `user disconnects from voice` have been disabled");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                case "levelannouncements":
-                    server.LogLevelUpAnnouncements = 0;
-                    Servers.SaveServers();
-                    await GlobalCommandResponses.CreateCommandResponse(Context,
-                        stopWatch.ElapsedMilliseconds,
-                        "Log Channel Set",
-                        $"{Context.User.Mention} Log messages for `level announcements` have been disabled.");
-                    break;
-                case "shadowbans":
-                    server.LogShadowbans = 0;
-                    Servers.SaveServers();
-                    await GlobalCommandResponses.CreateCommandResponse(Context,
-                        stopWatch.ElapsedMilliseconds,
-                        "Log Channel Set",
-                        $"{Context.User.Mention} Log messages for `shadowbans` have been disabled.");
-                    break;
-                case "unshadowbans":
-                    server.LogUnShadowbans = 0;
-                    Servers.SaveServers();
-                    await GlobalCommandResponses.CreateCommandResponse(Context,
-                        stopWatch.ElapsedMilliseconds,
-                        "Log Channel Set",
-                        $"{Context.User.Mention} Log messages for `un-shadowbans` have been disabled.");
-                    break;
-                case "all":
-                    server.LogDeletedMessages = 0;
-                    server.LogUpdatedMessages = 0;
-                    server.LogWhenUserJoins = 0;
-                    server.LogWhenUserLeaves = 0;
-                    server.LogBans = 0;
-                    server.LogUnbans = 0;
-                    server.LogChangesToLogSettings = 0;
-                    server.LogWhenUserSaysFilteredPhrase = 0;
-                    server.LogWhenUserConnectsToVoiceChannel = 0;
-                    server.LogWhenUserDisconnectsFromVoiceChannel = 0;
-                    server.LogLevelUpAnnouncements = 0;
-                    server.LogShadowbans = 0;
-                    server.LogUnShadowbans = 0;
-                    Servers.SaveServers();
-                    embed.WithTitle("Log Channel Reset");
-                    embed.WithDescription($"{Context.User.Mention} Log messages for `everything` have been disabled.");
-                    embed.WithColor(Pink);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-                default:
-                    embed.WithTitle("Invalid Log Specification");
-                    embed.WithDescription($"{Context.User.Mention} Please specify a valid logging type!");
-                    embed.WithColor(Red);
-                    await BE(); stopWatch.Stop();
-                    logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds); break;
-            }
-        }
-
-        [Command("logtypes")] //administration
-        [Alias("loglist")]
-        [RequireUserPermission(GuildPermission.Administrator)]
-        [RequireBotPermission(GuildPermission.Administrator)]
-        public async Task LogList()
-        {
-            stopWatch.Start();
-            string cmdPrefix = Servers.GetServer(Context.Guild).commandPrefix;
-            var server = Servers.GetServer(Context.Guild);
-            ulong[] logChannels = { server.LogDeletedMessages, server.LogUpdatedMessages, server.LogWhenUserJoins, server.LogWhenUserLeaves,
-            server.LogBans, server.LogUnbans, server.LogChangesToLogSettings, server.LogWhenUserSaysFilteredPhrase,
-            server.LogWhenUserConnectsToVoiceChannel, server.LogWhenUserDisconnectsFromVoiceChannel, server.LogLevelUpAnnouncements, server.LogShadowbans,
-            server.LogUnShadowbans};
-            embed.WithTitle("List of Log Events");
-            embed.WithDescription($"List of all types of logging you can subscribe to. Use these with {cmdPrefix}log to enable logging!" +
-                "\n" +
-                $"\n**DeletedMessages** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[0])}`**" +
-                $"\n**UpdatedMessages** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[1])}`**" +
-                $"\n**UserJoins** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[2])}`**" +
-                $"\n**UserLeaves** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[3])}`**" +
-                $"\n**Bans** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[4])}`**" +
-                $"\n**Unbans** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[5])}`**" +
-                $"\n**ChangesToLogSettings** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[6])}`**" +
-                $"\n**FilteredPhrases** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[7])}`**" +
-                $"\n**UserConnectsToVoice** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[8])}`**" +
-                $"\n**UserDisconnectsFromVoice** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[9])}`**" +
-                $"\n**LevelAnnouncements** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[10])}`**" +
-                $"\n**Shadowbans** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[11])}`**" +
-                $"\n**Unshadowbans** - Currently Assigned to: **`#{Context.Guild.GetChannel(logChannels[12])}`**" +
-                "\n**All**");
-            embed.WithColor(Pink);
-            await BE(); stopWatch.Stop();
-            logger.ConsoleCommandLog(Context, stopWatch.ElapsedMilliseconds);
-        }
-
     }
 }
